Move extensions filtering into selectors
