+++ src/components/Extensions/index.js
@@ -9,11 +9,11 @@ import Toolbar from "@material-ui/core/Toolbar";
 import MenuDrawer from "components/MenuDrawer";
 import RefreshButton from "components/RefreshButton";
 import ExtensionList from "components/Extensions/ExtensionList";
-import type { ExtensionType } from "types";
 import { useSelector, useDispatch } from "react-redux";
 import {
   selectIsExtensionsLoading,
-  selectExtensions,
+  selectInstalledExtensions,
+  selectNotInstalledExtensions,
   fetchExtensions,
   installExtension,
   uninstallExtension,
@@ -25,7 +25,8 @@ import {
 // but also because I don't think it's worth the effort to implement.
 
 const Extensions = () => {
-  const extensions = useSelector(selectExtensions);
+  const installedExtensions = useSelector(selectInstalledExtensions);
+  const notInstalledExtensions = useSelector(selectNotInstalledExtensions);
   const isExtensionsLoading = useSelector(selectIsExtensionsLoading);
 
   const dispatch = useDispatch();
@@ -41,27 +42,6 @@ const Extensions = () => {
   useEffect(() => {
     dispatch(fetchExtensions());
   }, []); // eslint-disable-line react-hooks/exhaustive-deps
-
-  function extensionSort(a: ExtensionType, b: ExtensionType) {
-    // First sort alphabetically by language
-    // Not using the pretty print / native name, but it gets the job done
-    if (a.lang > b.lang) return 1;
-    if (a.lang < b.lang) return -1;
-
-    // Then sort alphabetically by source name
-    if (a.name > b.name) return 1;
-    if (a.name < b.name) return -1;
-
-    return 0;
-  }
-
-  const installedExtensions = extensions
-    .filter(extension => extension.status === "INSTALLED")
-    .sort(extensionSort);
-
-  const notInstalledExtensions = extensions
-    .filter(extension => extension.status !== "INSTALLED")
-    .sort(extensionSort);
 
   return (
     <React.Fragment>
